tested offline messages now working fine
"0" in redis works for the messages which got dropped in server as a backup

">" this is for new message this is being used as delivery method for messages which are pushed but the other socket was offline.

"0" is for insurance. only heavft lifting is done via ">"

currently we are following the philosophy of exact once delivery method.

why ??? as we have done xack and xdel both once we recieve the chat ack

this could be made better delete only once the while sending the ack the other user is offline